[TS-385] Refactor: user type refactoring
