+++ src/components/common/Modal/CommonModal/SelectGenderModal/SelectGenderModal.tsx
@@ -8,7 +8,7 @@ import { genderTypeList } from "@/constants/onboarding";
 
 import { generateGender } from "@/utils/generateRangeType";
 
-import type { OnboardingUserInfoType } from "@/types/userDataType";
+import type { OnboardingUserInfoType } from "@/types/user";
 
 import {
 	layoutStyle,
